Revert "change binary-state-length to content-length"
This reverts commit 07a59d7e8223930100df878132b27ea819d253e5.

+++ kii-core/kii_core.h
@@ -564,7 +564,7 @@ kii_core_upload_thing_state(
         const char* state,
         const char* content_type,
         const char* content_encoding,
-        const int binary_body_length);
+        const size_t binary_body_length);
 
 /** prepare request of create object.
  * after this method succeeded, state of SDK becomes KII_STATE_READY.<br>
